test: move rebatch elision test into operators/rebatch.rs
streamling-core keeps unit tests inline in source modules rather than
standalone files under tests/; move the ProjectionPushdown regression
test into a #[cfg(test)] module next to RebatchExec.
